# docs/schemas/proxy_openapi.yaml
openapi: 3.1.0
info:
  title: ATLAS Proxy API
  description: >
    HTTP + SSE surface of atlas-proxy (the agent control plane). Versioned
    by GET /version (api_version); breaking changes bump the major. Error
    responses use the stable error envelope (see error_envelope.schema.json)
    — clients switch on the `error` code, never the human `detail`.
  version: 1.0.0
servers:
  - url: http://localhost:8090
    description: Local proxy (loopback only by default)
paths:
...
  /v1/agent:
    post:
      summary: Run one agent turn; streams SSE events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: {type: string}
                working_dir: {type: string}
                mode: {type: string, enum: [default, accept-edits, yolo]}
                session_id: {type: string}
                history:
                  type: array
                  description: Prior-turn messages replayed before the new one (most recent 40 kept)
                  items:
                    type: object
                    properties:
                      role: {type: string, enum: [user, assistant]}
                      content: {type: string}
                session_allowed_tools:
                  type: array
                  description: Tool names pre-approved for the session (skip the permission prompt)
                  items: {type: string}
                bypass_v3: {type: boolean, description: Disable V3 orchestration for the turn}
                disable_fresh_slot: {type: boolean, description: Keep the pre-warmed KV-cache prefix instead of requesting a fresh slot}
                sandbox_subdir: {type: string, description: Confine the turn to a bare subdirectory name under the workspace}
      responses:
        "200":
          description: SSE stream of events (see sse_envelope.schema.json)
          content:
            text/event-stream: {schema: {type: string}}
        "401": {$ref: "#/components/responses/Unauthorized"}
...
  /cancel:
    post:
      summary: Abort the in-flight turn for a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id]
              properties: {session_id: {type: string}}
      responses:
        "200": {description: Cancelled}
        "404": {description: No in-flight turn for that session (idempotent repeat)}
  /v1/permission:
    post:
      summary: Answer a pending permission_request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, tool_call_id, decision]
              properties:
                session_id: {type: string}
                tool_call_id: {type: string}
                decision: {type: string, enum: [allow, deny]}
                scope: {type: string, enum: [once, session]}
      responses:
        "200": {description: Decision delivered}
        "404": {description: No matching pending request (already resolved, cancelled, or timed out)}
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      description: >
        Per-installation service token (secrets/service-token). Required on
        all routes except /health, /ready, /version when a token is
        configured.
  responses:
    Unauthorized:
      description: Missing/invalid service token (error code `unauthorized`)
      content:
        application/json:
          schema: {$ref: "./error_envelope.schema.json"}
security:
  - serviceToken: []